chore(ci) debug e2e test for quicker feedback loop

+++ test/e2e/externalservices_kubernetes_test.go
@@ -148,6 +148,9 @@ metadata:
 	})
 
 	AfterEach(func() {
+		if ShouldSkipCleanup() {
+			return
+		}
 		err := cluster.DeleteKuma(deployOptsFuncs...)
 		Expect(err).ToNot(HaveOccurred())
